requirements.txt
gdown
tensorflow==2.14.0
streamlit
pillow
numpy
pandas